Refactor: Move class
