+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/list_view/ChatMessageListItem.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chat_messages;
+package bisq.desktop.main.content.components.chat_messages.list_view;
 
 import bisq.bonded_roles.market_price.MarketPriceService;
 import bisq.chat.ChatChannel;
