+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/list_view/ChatMessagesListController.java
@@ -33,7 +33,6 @@ import bisq.desktop.components.overlay.Popup;
 import bisq.desktop.main.content.bisq_easy.BisqEasyServiceUtil;
 import bisq.desktop.main.content.bisq_easy.take_offer.TakeOfferController;
 import bisq.desktop.main.content.components.ReportToModeratorWindow;
-import bisq.desktop.main.content.components.chat_messages.ChatMessageListItem;
 import bisq.i18n.Res;
 import bisq.network.NetworkService;
 import bisq.network.identity.NetworkId;
