+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/list_view/message_box/LeaveChatMessageBox.java
@@ -21,7 +21,7 @@ import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
 import bisq.desktop.common.utils.ImageUtil;
 import bisq.desktop.main.content.bisq_easy.open_trades.MyProtocolLogMessageBox;
-import bisq.desktop.main.content.components.chat_messages.ChatMessageListItem;
+import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessageListItem;
 import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessagesListController;
 import bisq.i18n.Res;
 import javafx.scene.control.Hyperlink;
